update and add credential.go
