+++ Soft_Eng_Research/Blockchain_Core/model/student.go
@@ -2,7 +2,6 @@ package blockchain
 
 import (
 	"bytes"
-	"crypto/sha256"
 	"fmt"
 	"time"
 )
@@ -14,19 +13,6 @@ type Student struct {
 	Age         int
 	BirthDate   time.Time
 	Credentials []Credential
-}
-
-type Credential struct {
-	Type       string
-	Issuer     string
-	DataIssued time.Time
-	Hash       []byte
-}
-
-// Serialize converts the Credential to a byte array
-func (cred Credential) Serialize() []byte {
-	// Convert the Credential to a byte array
-	return []byte(fmt.Sprintf("%s|%s|%s", cred.Type, cred.Issuer, cred.DataIssued.Format(time.RFC3339)))
 }
 
 // AddNewStudent creates and returns a new student
@@ -95,16 +81,6 @@ func FindStudentByID(id int) (*Student, error) {
 	return nil, fmt.Errorf("Student not found")
 }
 
-// VerifyCredential checks if a credential is valid by comparing its hash
-func VerifyCredential(s *Student, cred Credential) bool {
-	for _, storedCred := range s.Credentials {
-		if bytes.Equal(storedCred.Hash, cred.Hash) {
-			return true
-		}
-	}
-	return false
-}
-
 // DeleteCredential removes a credential from the student's list
 func DeleteCredential(s *Student, cred Credential) {
 	for i, storedCred := range s.Credentials {
@@ -114,37 +90,6 @@ func DeleteCredential(s *Student, cred Credential) {
 			break
 		}
 	}
-}
-
-// GenerateCredentialHash creates a hash for a credential
-func GenerateCredentialHash(cred Credential) []byte {
-	// Combine the credential fields to generate a unique hash
-	credData := fmt.Sprintf("%s%s%s", cred.Type, cred.Issuer, cred.DataIssued.String())
-	hash := sha256.Sum256([]byte(credData)) // Using SHA-256 to hash the credential data
-	return hash[:]
-}
-
-// ValidateCredentialData validates the data of the credential (to be implemented)
-func ValidateCredentialData(cred Credential) error {
-	// Add validation logic for credential data if needed
-	// in this case, we can check if the credential type is valid
-	if cred.Type == "" {
-		return fmt.Errorf("credential type cannot be empty")
-	}
-
-	// Check if Issuer is empty
-	if cred.Issuer == "" {
-		return fmt.Errorf("issuer cannot be empty")
-	}
-
-	// Check if DataIssued is a future date
-	if cred.DataIssued.After(time.Now()) {
-		return fmt.Errorf("issued date cannot be in the future")
-	}
-
-	// Additional validation checks can be added here as needed
-
-	return nil
 }
 
 // LookupCredentials retrieves and displays all the credentials of the student
